chore: integrate changes required by Theia
Adapts the "menubar" to support documents different than the global
one. This is required when rendering a menubar in a different window
which has its own document.

Adapts the "menu" to remove the check for empty menus. Menus in Theia
are dynamic and therefore by default empty until they are shown. This
check therefore prevents opening all menus in Theia.